Split the rpc_agent module into multiple submodules
