+++ src/agentscope/agents/__init__.py
@@ -6,7 +6,7 @@ from .dialog_agent import DialogAgent
 from .dict_dialog_agent import DictDialogAgent
 from .user_agent import UserAgent
 from .text_to_image_agent import TextToImageAgent
-from .rpc_agent import RpcAgent, RpcAgentServerLauncher
+from .rpc_agent import RpcAgent
 from .react_agent import ReActAgent
 
 
@@ -20,5 +20,4 @@ __all__ = [
     "ReActAgent",
     "DistConf",
     "RpcAgent",
-    "RpcAgentServerLauncher",
 ]
